% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clean_rob.R
\name{create_clean_rob}
\alias{create_clean_rob}
\title{Cleans rob dataframe}
\usage{
create_clean_rob(rob)
}
\arguments{
\item{rob}{Risk of bias export from covidence}
}
\value{

}
\description{
Cleans rob dataframe
}
\author{
Aaron Conway
}
