% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfmake_characteristics_table.R,
%   R/make_characteristics_table.R
\name{make_characteristics_table}
\alias{make_characteristics_table}
\title{Table of study characteristics}
\usage{
make_characteristics_table(study_df, baseline_df, outcomes_df)

make_characteristics_table(study_df, baseline_df, outcomes_df)
}
\arguments{
\item{study_df}{data frame}

\item{baseline_df}{data frame}

\item{outcomes_df}{data frame}
}
\value{



}
\description{
Table of study characteristics

Table of study characteristics
}
