% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ma_df.R
\name{make_ma_df}
\alias{make_ma_df}
\title{Preprocessing for meta-analysis}
\usage{
make_ma_df(outcomes_df = outcomes_df, rob = rob, study_df = study_df, ...)
}
\arguments{
\item{outcomes_df}{Filtered outcome_df with selection of studies}

\item{rob}{Data frame of risk of bias assessments}

\item{study_df}{Data frame of study characteristics}
}
\description{
Make a data frame for analyses
}
