% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prisma.R
\name{make_prisma}
\alias{make_prisma}
\title{Make a prisma diagram}
\usage{
make_prisma(
  retrieved,
  duplicates,
  included,
  full_text,
  wrong_setting,
  wrong_population,
  wrong_outcome,
  line_size = 0.5,
  box_type = "round",
  box_lines = 0.5,
  box_colour = "#ff4f14",
  box_fill = "white",
  text_colour = "#ff4f14",
  arrow_colour = "#ff4f14"
)
}
\value{

}
\description{
Make a prisma diagram
}
