% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_flextable.R
\name{make_results_flextable}
\alias{make_results_flextable}
\title{Make a table of the meta-analyses}
\usage{
make_results_flextable(results_list, names)
}
\arguments{
\item{results_list}{list of meta-analysis results}
}
\value{

}
\description{
Make a table of the meta-analyses
}
\author{
Aaron Conway
}
