% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rob_across.R
\name{make_rob_across}
\alias{make_rob_across}
\title{Makes a figure of risk of bias across studies}
\usage{
make_rob_across(rob)
}
\value{

}
\description{
Makes a figure of risk of bias across studies
}
