% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rob_within.R
\name{make_rob_within}
\alias{make_rob_within}
\title{Makes a figure of risk of bias within studies}
\usage{
make_rob_within(rob)
}
\value{

}
\description{
Makes a figure of risk of bias within studies
}
