% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Data frame of meta-analysis results}
\usage{
meta_analysis(data)
}
\arguments{
\item{data}{Data frame of subsetted studies}
}
\value{

}
\description{
Creates meta-anlysis results
}
